Implement PreSigned Url
We generated a Presigned Url for PUT and GET

PUT Presigned url will be used to make a put request to upload an image to our S3 bucket

GET presinged url will be used to load an image

Both links expires after 5 mins

FYI: I did some code refactoring in for our multiplexer. Route matching with the julienschmidt multiplexer was so bad!
https://github.com/gin-gonic/gin/issues/1301